pyproject.toml
[project]
name = "politrack"
version = "0.1.0"
description = "A Hierarchical Political Management System for assigning tasks, managing user roles, and enabling communication across multiple administrative layers. Built with FastAPI, MongoDB, and Jinja2 with role-based access and real-time interaction features."
authors = [
    { name = "Sunny Raj", email = "mail.sunnyraj94@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"

dependencies = [
    'annotated-types==0.7.0',
    'anyio==4.9.0',
    'certifi==2025.6.15',
    'cffi==1.17.1',
    'click==8.2.1',
    'cryptography==45.0.5',
    'dnspython==2.7.0',
    'ecdsa==0.19.1',
    'email_validator==2.2.0',
    'fastapi==0.115.14',
    'gunicorn==23.0.0',
    'h11==0.16.0',
    'httpcore==1.0.9',
    'httpx==0.27.0',
    'idna==3.10',
    'Jinja2==3.1.4',
    'MarkupSafe==3.0.2',
    'packaging==25.0',
    'passlib==1.7.4',
    'pyasn1==0.6.1',
    'pycparser==2.22',
    'pydantic==2.11.7',
    'pydantic_core==2.33.2',
    'pymongo==4.13.2',
    'python-dotenv==1.1.1',
    'python-jose==3.5.0',
    'rsa==4.9.1',
    'setuptools==78.1.1',
    'six==1.17.0',
    'sniffio==1.3.1',
    'starlette==0.46.2',
    'typing-inspection==0.4.1',
    'typing_extensions==4.14.0',
    'uvicorn==0.35.0',
    'wheel==0.45.1',
    'beanie==1.30.0',
    'fastapi_users==14.0.1',
    'fastapi-users[beanie]'
    ]

[tool.setuptools.packages.find]
where = ["app"]
include = [
    "app*",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
